@@ -15,18 +15,22 @@
     "@babel/core": "^7.11.4",
     "@babel/plugin-proposal-class-properties": "^7.10.4",
     "@babel/plugin-transform-react-jsx": "^7.10.4",
-    "@babel/preset-react": "^7.0.0",
+    "@babel/preset-react": "^7.16.7",
     "@babel/preset-typescript": "^7.12.1",
-    "@emotion/core": "^10.0.35",
-    "@emotion/styled": "^10.0.27",
+    "@emotion/core": "^11.0.0",
+    "@emotion/eslint-plugin": "^11.7.0",
+    "@emotion/styled": "^11.6.0",
     "@svgr/cli": "^5.4.0",
     "@testing-library/dom": "^7.1.2",
-    "@types/react": "^16.9.55",
+    "@types/react": "^16.14.21",
     "@types/react-dom": "^16.9.9",
+    "@types/react-modal": "^3.13.1",
     "@types/react-redux": "^7.1.9",
     "@types/react-router-dom": "^5.1.6",
+    "@types/react-transition-group": "^4.4.4",
+    "@types/uuid": "^8.3.4",
     "axios": "^0.19",
-    "babel-plugin-macros": "^2.8.0",
+    "babel-plugin-macros": "^3.1.0",
     "browser-sync": "^2.26.7",
     "browser-sync-webpack-plugin": "^2.0.1",
     "concurrently": "^5.3.0",
@@ -34,26 +38,31 @@
     "laravel-mix": "^5.0.1",
     "laravel-mix-react-typescript-extension": "^1.0.3",
     "prettier": "^2.1.2",
-    "react": "^16.13.1",
+    "react": "^16.14.0",
     "react-dom": "^16.13.1",
+    "redux-devtools-extension": "^2.13.9",
     "remove-files-webpack-plugin": "^1.4.0",
     "resolve-url-loader": "^3.1.0",
     "sass": "^1.26.3",
     "sass-loader": "^8.0.2",
     "schemats": "^3.0.3",
-    "twin.macro": "^1.7.0",
+    "twin.macro": "^2.8.2",
     "typescript": "^4.0.5"
   },
   "dependencies": {
     "@emotion/babel-plugin-jsx-pragmatic": "^0.1.5",
+    "@emotion/babel-preset-css-prop": "^11.2.0",
+    "@emotion/react": "^11.7.1",
     "@react-hook/window-size": "^3.0.7",
     "D": "^1.0.0",
     "axios-progress-bar": "git+https://github.com/HynekS/progress-bar-4-axios-updated.git#master",
     "babel-plugin-transform-react-jsx": "^6.24.1",
     "babel-plugin-transform-react-jsx-source": "^6.22.0",
+    "babel-plugin-twin": "^1.0.2",
     "date-fns": "^2.28.0",
     "dotenv": "^14.1.0",
     "filesize.js": "^2.0.0",
+    "heroicons": "^1.0.5",
     "js-file-download": "^0.4.11",
     "laravel-mix-bundle-analyzer": "^1.0.5",
     "lodash.deburr": "^4.1.0",
@@ -67,22 +76,18 @@
     "react-highlight-words": "^0.16.0",
     "react-hook-form": "^5.7.2",
     "react-lazy-with-preload": "^2.0.1",
-    "react-modal": "^3.11.2",
+    "react-modal": "^3.14.4",
     "react-redux": "^7.2.0",
-    "react-router": "^6.0.0-alpha.2",
-    "react-router-dom": "^6.0.0-alpha.2",
-    "react-transition-group": "^4.3.0",
+    "react-router": "^6.2.1",
+    "react-router-dom": "^6.2.1",
+    "react-transition-group": "^4.4.2",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
-    "tailwindcss": "^1.7.5"
+    "tailwindcss": "^3.0.15"
   },
   "babelMacros": {
     "twin": {
-      "config": "tailwind.config.js",
-      "preset": "emotion",
-      "debugProp": true,
-      "debugPlugins": false,
-      "debug": false
+      "preset": "emotion"
     }
   }
 }
