@@ -1,5 +1,4 @@
-// changed from es6 import to supress TS errors:
-const { default: axios } = require("axios")
+import axios from "axios"
 
 import store from "../store/configuredStore"
 import { clearLoggedInUser } from "../store/auth"
@@ -13,7 +12,7 @@ axios.defaults.baseURL = API_URL
 axios.defaults.headers.common.Accept = "application/json"
 axios.defaults.headers.common["X-CSRF-TOKEN"] = document.querySelector<HTMLMetaElement>(
   'meta[name="csrf-token"]',
-).content
+)?.content
 axios.defaults.headers.common["X-Requested-With"] = "XMLHttpRequest"
 
 axios.interceptors.response.use(
