@@ -1,8 +1,5 @@
-/** @jsx jsx */
-import React from "react"
 import { useNavigate } from "react-router-dom"
 import { useSelector, useDispatch } from "react-redux"
-import { jsx } from "@emotion/core"
 import tw from "twin.macro"
 
 import { logout } from "../../store/auth"
@@ -10,7 +7,7 @@ import SvgLogout from "../../vendor/heroicons/outline/Logout"
 
 import type { AppState } from "../../store/rootReducer"
 
-const Button = tw.button`flex items-center bg-blue-500 hover:bg-blue-700 transition-colors duration-300 text-white font-medium py-2 px-4 rounded focus:(outline-none shadow-outline)`
+const Button = tw.button`flex items-center bg-blue-500 hover:bg-blue-700 transition-colors duration-300 text-white font-medium py-2 px-4 rounded focus:(outline-none ring)`
 
 const Logout = () => {
   const dispatch = useDispatch()
