@@ -1,8 +1,26 @@
-import React from "react"
-
 import SvgExclamationCircle from "../../vendor/heroicons/solid/ExclamationCircle"
 
-const Input = ({ name, label, register, error={}, placeholder = "", type = "text", ...props }) => {
+type Props = {
+  name: string
+  label: string
+  // Hotfix! Possibly faullty. Try to pluck the type from useForm
+  register: React.RefObject<HTMLInputElement> | React.LegacyRef<HTMLInputElement>
+  error?: {
+    [name: string]: Error
+  }
+  placeholder?: string
+  type: React.HTMLInputTypeAttribute
+}
+
+const Input = ({
+  name,
+  label,
+  register,
+  error = {},
+  placeholder = "",
+  type = "text",
+  ...props
+}: Props) => {
   return (
     <div className="fieldWrapper">
       <div className="labelWrapper">
@@ -18,7 +36,12 @@ const Input = ({ name, label, register, error={}, placeholder = "", type = "text
           className={error[name] ? "hasError" : ""}
           {...props}
         />
-        {error[name] && <div className="errorMessage"><SvgExclamationCircle/>{error[name].message}</div>}
+        {error[name] && (
+          <div className="errorMessage">
+            <SvgExclamationCircle />
+            {error[name].message}
+          </div>
+        )}
       </div>
     </div>
   )
