@@ -1,7 +1,5 @@
-// @ts-check
-/** @jsx jsx */
 import React from "react"
-import { jsx } from "@emotion/core"
+import { jsx } from "@emotion/react"
 import tw from "twin.macro"
 
 // Todo move DetailPage out of project (it is a layout stuff)
@@ -14,13 +12,13 @@ const Dashboard = () => {
         <h1 tw="text-xl">Dashboard</h1>
         <section tw="pb-8">
           <form action="">
-            <label htmlFor="quicksearch" tw="block text-gray-700 text-sm font-bold mb-2">
+            <label htmlFor="quicksearch" tw="block mb-2 text-sm font-bold text-gray-700">
               rychlé hledání
             </label>
             <input
               type="text"
               name="quicksearch"
-              tw="bg-gray-200 appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline focus:transition-shadow focus:duration-300"
+              tw="w-full px-3 py-2 leading-tight text-gray-700 bg-gray-200 border rounded appearance-none focus:outline-none focus:ring focus:transition-shadow focus:duration-300"
             />
           </form>
         </section>
