@@ -1,4 +1,3 @@
-// @ts-check
 import React from "react"
 import { useSelector, useDispatch } from "react-redux"
 import { useForm } from "react-hook-form"
@@ -12,8 +11,8 @@ import useFocusNextOnEnter from "../../hooks/useFocusNextOnEnter"
 const FormContainer = tw.div`w-full max-w-xs relative z-10`
 const Form = tw.form`bg-white shadow-xl rounded px-10 pt-8 pb-10 mb-4`
 const Label = tw.label`block text-gray-700 text-sm font-bold mb-2`
-const Input = tw.input`bg-gray-200 appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline focus:transition-shadow focus:duration-300`
-const Submit = tw.button`bg-blue-600 hover:bg-blue-700 transition-colors duration-300 text-white w-full font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline focus:transition-shadow focus:duration-300`
+const Input = tw.input`bg-gray-200 appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:ring focus:transition-shadow focus:duration-300`
+const Submit = tw.button`bg-blue-600 hover:bg-blue-700 transition-colors duration-300 text-white w-full font-bold py-2 px-4 rounded focus:outline-none focus:ring focus:transition-shadow focus:duration-300`
 const ErrorMessage = tw.div`text-red-600 text-sm pb-2`
 
 const LoginForm = () => {
