@@ -1,6 +1,5 @@
-/** @jsx jsx */
 import React from "react"
-import { jsx } from "@emotion/core"
+import { jsx } from "@emotion/react"
 import tw from "twin.macro"
 
 const DetailPage = ({ children }) => (
