@@ -1,6 +1,4 @@
-import * as React from "react"
-
-function DropFiles(props) {
+function DropFiles(props: unknown): JSX.Element {
   return (
     <svg fill="none" viewBox="0 0 420 421" {...props}>
       <path d="M266.714 407h-30a7 7 0 100 14h30a7 7 0 100-14zM210.714 407H183c-.561 0-1.124-.024-1.675-.072-3.864-.33-7.246 2.516-7.581 6.367a7 7 0 006.367 7.58c.953.083 1.925.125 2.889.125h27.714a7 7 0 100-14zM322.714 407h-30a7 7 0 100 14h30a7 7 0 100-14zM378.714 407h-30a7 7 0 100 14h30a7 7 0 100-14zM295 159h30a7 7 0 100-14h-30a7 7 0 100 14zM157 351.572a7 7 0 007-7v-30a7 7 0 10-14 0v30a7 7 0 007 7zM164 388v-17.428a7 7 0 10-14 0V388a33.096 33.096 0 003.778 15.344 6.999 6.999 0 009.453 2.936 7 7 0 002.937-9.454A18.781 18.781 0 01164 388zM413 206.143a7 7 0 00-7 7v30a7 7 0 1014 0v-30a7 7 0 00-7-7zM413 262.144a7 7 0 00-7 7v30a7 7 0 1014 0v-30a7 7 0 00-7-7zM409.947 154.285a7 7 0 00-9.736 10.06c3.732 3.613 5.788 8.463 5.789 13.655v9.143a7 7 0 1014 0v-9.144c-.001-9.019-3.571-17.44-10.053-23.714zM413 374.144a7 7 0 00-7 7V388a18.94 18.94 0 01-7.041 14.766 6.999 6.999 0 00-1.025 9.846 6.984 6.984 0 005.439 2.59c1.55 0 3.109-.512 4.406-1.564C415.546 407.336 420 397.991 420 388v-6.856a7 7 0 00-7-7zM413 318.144a7 7 0 00-7 7v30a7 7 0 1014 0v-30a7 7 0 00-7-7zM381 145h-30a7 7 0 100 14h30a7 7 0 100-14z" />
