@@ -1,7 +1,5 @@
-// @ts-check
-/** @jsx jsx */
 import React, { Fragment, useState } from "react"
-import { jsx, css } from "@emotion/core"
+import { jsx, css } from "@emotion/react"
 import tw from "twin.macro"
 
 import { modalStatus } from "./InvoiceModalStatus"
@@ -19,19 +17,19 @@ const Invoice = ({ invoice, modalOpenCallback }) => {
   return (
     <Fragment>
       <tr key={invoice.id_zaznam}>
-        <td tw="text-right px-2 py-2">{invoice.c_faktury}</td>
-        <td tw="text-right px-2 py-2">
+        <td tw="px-2 py-2 text-right">{invoice.c_faktury}</td>
+        <td tw="px-2 py-2 text-right">
           {new Date(Date.parse(invoice.datum_vlozeni)).toLocaleDateString("cs-CZ")}
         </td>
-        <td tw="text-right px-2 py-2">{invoice.castka.toLocaleString("cs-CZ")},–</td>
-        <td tw="text-right px-2 py-2">
+        <td tw="px-2 py-2 text-right">{invoice.castka.toLocaleString("cs-CZ")},–</td>
+        <td tw="px-2 py-2 text-right">
           <div ref={innerRef} tw="relative">
             <button tw="flex items-center" onClick={() => setIsMenuOpen(true)}>
               <SvgDotsHorizontalSolid tw="flex w-5 opacity-50" />
             </button>
             <div
               css={[
-                tw`absolute top-full left-0 text-sm bg-white shadow invisible rounded z-10`,
+                tw`absolute left-0 z-10 invisible text-sm bg-white rounded shadow top-full`,
                 isMenuOpen && tw`visible`,
               ]}
             >
@@ -39,14 +37,14 @@ const Invoice = ({ invoice, modalOpenCallback }) => {
                 tw="flex items-center w-full p-2 pr-4 rounded rounded-b-none focus:(outline-none) hocus:(bg-gray-200) transition-colors duration-300"
                 onClick={() => modalOpenCallback({ status: modalStatus.DESTROY, data: invoice })}
               >
-                <SvgTrashSolid tw="flex w-5 opacity-50 mr-2" />
+                <SvgTrashSolid tw="flex w-5 mr-2 opacity-50" />
                 Odstranit
               </button>
               <button
                 tw="flex items-center w-full p-2 pr-4 rounded rounded-t-none focus:(outline-none) hocus:(bg-gray-200) transition-colors duration-300"
                 onClick={() => modalOpenCallback({ status: modalStatus.UPDATE, data: invoice })}
               >
-                <SvgPencilSolid tw="flex w-5 opacity-50 mr-2" />
+                <SvgPencilSolid tw="flex w-5 mr-2 opacity-50" />
                 Upravit
               </button>
             </div>
