@@ -1,8 +1,6 @@
-// @ts-check
-/** @jsx jsx */
 import React, { useEffect } from "react"
 import { useForm } from "react-hook-form"
-import { jsx, css } from "@emotion/core"
+import { jsx, css } from "@emotion/react"
 import tw from "twin.macro"
 import { useSelector, useDispatch } from "react-redux"
 
@@ -31,7 +29,7 @@ const styles = css`
     & .DayPickerInput {
       position: relative;
       &::after {
-        ${tw`fill-current h-full w-4 mr-2 absolute inset-y-0 right-0 opacity-25`}
+        ${tw`absolute inset-y-0 right-0 w-4 h-full mr-2 opacity-25 fill-current`}
         content: "";
         background: url(/images/calendar-solid.svg) no-repeat center;
       }
@@ -41,7 +39,7 @@ const styles = css`
       ${tw`bg-gray-200 appearance-none border-2 border-gray-200 rounded py-1 px-2 text-gray-700 leading-tight focus:(outline-none bg-white border-blue-500)`}
     }
     & input[type="checkbox"] {
-      ${tw`w-auto focus:(outline-none shadow-outline)`}
+      ${tw`w-auto focus:(outline-none ring)`}
     }
     & select {
       ${tw`block appearance-none border-2 w-full bg-white border-gray-300 py-1 pl-2 pr-8 rounded leading-tight hover:(border-gray-400) focus:(outline-none bg-white border-blue-500)`}
@@ -55,7 +53,7 @@ const styles = css`
     }
   }
   .errorMessage {
-    ${tw`absolute flex z-10 left-0 top-full text-xs text-red-600 bg-white p-1 pr-2 rounded shadow-sm`}
+    ${tw`absolute left-0 z-10 flex p-1 pr-2 text-xs text-red-600 bg-white rounded shadow-sm top-full`}
     & svg {
       ${tw`w-4 mr-2 fill-current`}
     }
@@ -116,15 +114,15 @@ const InvoiceUpdateForm = ({ modalState: { data }, onModalClose, ...props }) => 
             error={errors}
           />
         </div>
-        <footer tw="flex justify-end bg-gray-100 p-6 rounded-lg rounded-t-none">
+        <footer tw="flex justify-end p-6 bg-gray-100 rounded-lg rounded-t-none">
           <button
-            tw="text-gray-500 font-medium py-2 px-4 ml-4 rounded transition-colors duration-300 hover:(text-gray-600) focus:(outline-none shadow-outline transition-shadow duration-300)"
+            tw="text-gray-500 font-medium py-2 px-4 ml-4 rounded transition-colors duration-300 hover:(text-gray-600) focus:(outline-none ring transition-shadow duration-300)"
             onClick={onModalClose}
           >
             Zrušit
           </button>
           <button
-            tw="bg-blue-600 transition-colors duration-300 text-white font-medium py-2 px-4 ml-4 rounded hover:(bg-blue-700) focus:(outline-none shadow-outline transition-shadow duration-300)"
+            tw="bg-blue-600 transition-colors duration-300 text-white font-medium py-2 px-4 ml-4 rounded hover:(bg-blue-700) focus:(outline-none ring transition-shadow duration-300)"
             className={`${isLoading ? "spinner" : ""}`}
             type="submit"
           >
