@@ -1,4 +1,4 @@
-export default function getImageOrFallback(path, fallback) {
+export default function getImageOrFallback(path: string, fallback: string) {
   return new Promise(resolve => {
     const img = new Image()
     img.src = path
